stt/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
openai-whisper>=20231117
requests>=2.31.0